Add vanilla CSS support to Toolbar (#505)
* Add vanilla CSS support to Toolbar

* PR feedback
