+++ data/primitives/components/toolbar/1.0.1.mdx
@@ -28,9 +28,7 @@ aria: https://www.w3.org/WAI/ARIA/apg/patterns/toolbar
   <Box css={{ flex: '0 0 20px', '@bp1': { flex: '0 0 60px' } }} />
 </HeroContainer>
 
-```jsx hero template=Toolbar
-
-```
+<HeroCodeBlock folder="Toolbar" />
 
 <Highlights features={['Full keyboard navigation.']} />
